refactoring message handling into a trait, first phase
