+++ src/lib.rs
@@ -20,21 +20,23 @@ use std::os::raw::c_uint;
 
 #[no_mangle]
 pub unsafe extern "C" fn c_get_next_stream_id() -> c_uint {
-    get_next_stream_id()
+  get_next_stream_id()
 }
 
 pub fn get_next_stream_id() -> u32 {
-    NEXT_EVENT_ID.fetch_add(1, std::sync::atomic::Ordering::SeqCst) as u32
+  NEXT_EVENT_ID.fetch_add(1, std::sync::atomic::Ordering::SeqCst) as u32
 }
 
 pub fn build_number() -> &'static str {
   match build_info::GIT_VERSION {
     Some(v) => v,
-    None => "unknown"
+    None => "unknown",
   }
 }
 
 pub mod build_info;
+
+pub mod js;
 
 pub mod errors;
 pub mod msg;
@@ -43,11 +45,11 @@ pub mod runtime;
 
 pub mod utils;
 
+pub mod acme_store;
 pub mod cache_store;
 pub mod cache_store_notifier;
 pub mod data_store;
 pub mod fs_store;
-pub mod acme_store;
 
 pub mod settings;
 
@@ -61,14 +63,16 @@ pub mod http_server;
 pub mod metrics;
 
 pub mod module_resolver;
-pub use crate::module_resolver::{ ModuleResolver, JsonSecretsResolver, LocalDiskModuleResolver };
+pub use crate::module_resolver::{JsonSecretsResolver, LocalDiskModuleResolver, ModuleResolver};
+
+pub mod msg_handler;
 
 mod disk_fs;
 mod postgres_data;
+mod redis_acme;
 mod redis_cache;
 mod redis_cache_notifier;
 mod redis_fs;
-mod redis_acme;
 mod sqlite_cache;
 mod sqlite_data;
