+++ src/ops/fetch.rs
@@ -6,7 +6,8 @@ use futures::{
 use crate::msg;
 use flatbuffers::FlatBufferBuilder;
 
-use crate::runtime::{JsBody, JsHttpResponse, JsRuntime, Op, EVENT_LOOP};
+use crate::js::*;
+use crate::runtime::{JsRuntime, EVENT_LOOP};
 use crate::utils::*;
 use libfly::*;
