+++ src/ops/image.rs
@@ -6,7 +6,8 @@ use libc::{c_float, c_int, c_void};
 use crate::msg;
 use flatbuffers::FlatBufferBuilder;
 
-use crate::runtime::{JsBody, JsRuntime, Op};
+use crate::js::*;
+use crate::runtime::JsRuntime;
 use crate::utils::*;
 use libfly::*;
