+++ src/utils.rs
@@ -4,20 +4,35 @@ use libfly::*;
 
 use crate::errors::FlyError;
 
-use crate::runtime::{Buf, JsBody, JsRuntime, Op};
+use crate::runtime::JsRuntime;
 
-use futures::{future, Future, Stream};
+use futures::{
+  future,
+  sync::{mpsc, oneshot},
+  Future, Stream,
+};
 
+use crate::js::*;
 use std::ptr;
+
+// Buf represents a byte array returned from a "Op".
+// The message might be empty (which will be translated into a null object on
+// the javascript side) or it is a heap allocated opaque sequence of bytes.
+// Usually a flatbuffer message.
+pub type Buf = Option<Box<[u8]>>;
+
+// JS promises in Fly map onto a specific Future
+// which yields either a FlyError or a byte array.
+pub type Op = Future<Item = Buf, Error = FlyError> + Send;
+pub type Handler = fn(JsRuntime, &msg::Base, fly_buf) -> Box<Op>;
 
 pub fn take_last_n(str: &str, n: usize) -> Option<&str> {
   if str.len() >= n {
-    Some(&str[str.len()-n..])
+    Some(&str[str.len() - n..])
   } else {
     None
   }
 }
-
 
 pub fn serialize_response(
   cmd_id: u32,
@@ -83,7 +98,8 @@ pub fn send_body_stream(ptr: JsRuntime, req_id: u32, stream: JsBody) {
         .for_each(move |v| {
           send_stream_chunk(ptr, req_id, v.as_ptr() as *mut u8, v.len(), false);
           Ok(())
-        }).and_then(move |_| {
+        })
+        .and_then(move |_| {
           send_done_stream(ptr, req_id);
           Ok(())
         }),
@@ -100,7 +116,8 @@ pub fn send_body_stream(ptr: JsRuntime, req_id: u32, stream: JsBody) {
           .for_each(move |v| {
             send_stream_chunk(ptr, req_id, v.as_ptr() as *mut u8, v.len(), false);
             Ok(())
-          }).and_then(move |_| {
+          })
+          .and_then(move |_| {
             send_done_stream(ptr, req_id);
             Ok(())
           }),
@@ -112,7 +129,8 @@ pub fn send_body_stream(ptr: JsRuntime, req_id: u32, stream: JsBody) {
           .for_each(move |mut b| {
             send_stream_chunk(ptr, req_id, b.as_mut_ptr() as *mut u8, b.len(), false);
             Ok(())
-          }).and_then(move |_| {
+          })
+          .and_then(move |_| {
             send_done_stream(ptr, req_id);
             Ok(())
           }),
@@ -131,7 +149,8 @@ pub fn send_body_stream(ptr: JsRuntime, req_id: u32, stream: JsBody) {
               false,
             );
             Ok(())
-          }).and_then(move |_| {
+          })
+          .and_then(move |_| {
             send_done_stream(ptr, req_id);
             Ok(())
           }),
@@ -159,7 +178,8 @@ pub fn send_stream_chunk(ptr: JsRuntime, req_id: u32, chunk: *mut u8, len: usize
           msg_type: msg::Any::StreamChunk,
           ..Default::default()
         },
-      ).unwrap(),
+      )
+      .unwrap(),
     ),
     Some(fly_buf {
       alloc_ptr: ptr::null_mut() as *mut u8,
@@ -189,8 +209,21 @@ pub fn send_done_stream(ptr: JsRuntime, req_id: u32) {
           msg_type: msg::Any::StreamChunk,
           ..Default::default()
         },
-      ).unwrap(),
+      )
+      .unwrap(),
     ),
     None,
   );
+}
+
+pub enum EventResponseChannel {
+  Http(oneshot::Receiver<JsHttpResponse>),
+  Dns(oneshot::Receiver<JsDnsResponse>),
+}
+
+#[derive(Debug)]
+pub enum EventDispatchError {
+  PoisonedLock,
+  Http(mpsc::SendError<JsHttpRequest>),
+  Dns(mpsc::SendError<JsDnsRequest>),
 }
